Replace a facility: crc32() with crc
Reason: see below.

This is a redesign of our CRC component. The function template crc32(),
specific for CRC-32, has been replaced with a traits-based class
template.

Advantages of the new design are:

 - it allows calculating any kind of known (and unknown) CRCs, while
   only moderately complicating the implementation

 - it allows calculating a CRC in multiple steps, for applications that
   don't have the whole set of data available at once

 - it doesn't need an external generator for the lookup table (cache):
   the table is now a constexpr object generated during compilation

Note that the implementation is a bit different in style:

 - it uses noexcept liberally (though I'm not sure that's a good idea,
   as the compiler sees the whole body of all the functions anyway)

 - it doesn't use std::accumulate() for the sake of using it

Note, too, that crc_test.cpp is a rename of crc32_test.cpp, with a few
modifications, but Git doesn't show it as a rename.