Prepare the Doxygen markup for re-enabling Markdown support
Reason: the Markdown syntax is cleaner than the equivalent Doxygen
syntax (cfr. e.g. backticks with \c or \code), so, if we can use it,
it's probably better, especially when the documentation is read directly
from the source files.

The fix, here, is to increase the indentation of some commands, to avoid
that the text that follows them is rendered as code because it's more
indented with respect to the commands.

Note that, at the time I'm writing this (October 27, 2020), DocsForge
builds with Markdown support enabled, so this fixes the DocsForge
output, too.