Improve the implementation of split_if()
This avoids the 'if' within the 'while' and gives the 'end' variable
consistent semantics (i.e., the same outside the 'while'). Of course, we
needed to do an 'emplace_back()' outside of the loop, but that's just
like before (except that we do it at the start of the function, now, not
at the end).